Conflicts:
	VisualC++/Quadra.vcproj
	VisualC++/Skelton.vcproj
	VisualC++/wadder.vcproj
	skelton/common/http_request.cpp
	skelton/include/net_buf.h
	skelton/svgalib/main.cpp
	skelton/svgalib/video_x11.cpp
	skelton/svgalib/video_x11_16.cpp
	skelton/svgalib/video_x11_24.cpp
	skelton/svgalib/video_x11_8.cpp
	source/quadra.cpp

+++ skelton/common/random.cpp
@@ -36,14 +36,14 @@ Random::Random(time_t p) {
 Word Random::rnd(int _and) { // better proc
 	int tmp;
 	seed = seed * 0x41c64e6d + 0x00003039;
-	tmp = seed >> 10;
+	tmp = int(seed >> 10);
 
 	return (Word) (tmp & _and);
 };
 
 Word Random::crap_rnd(int _and) { // crappy proc
 	int tmp;
-	tmp = seed * 0x41c64e6d + 0x00003039;
+	tmp = int(seed * 0x41c64e6d + 0x00003039);
 	seed = tmp >> 10;
 
 	return (Word) (tmp & _and);
